Fix the problem with crashes when rotating the device while asynctask are running in the background.
The problem occured because the Activity is recreated when the device is
turned. The fix is to store a reference to the AsyncTask in the application
object, and to reattach to the task when the activity is recreated.

Inspiration and code taken from www.fattybeagle.com/2011/02/15/android-asynctasks-during-a-screen-rotation-part-ii/ .

Now a reference to the AsyncTask is stored in the application object (which is not recreated when